[spirv] refactoring for better code structure
* Moved BlockReadableOrder.h, DeclResultIdMapper.h, and
  TypeTranslator.h into lib/SPIRV so that they are lib internal
  headers
* Extracted the SPIRVEmitter class into its own files

No logic change at all in this commit
