+++ tools/clang/lib/SPIRV/Structure.cpp
@@ -9,7 +9,7 @@
 
 #include "clang/SPIRV/Structure.h"
 
-#include "clang/SPIRV/BlockReadableOrder.h"
+#include "BlockReadableOrder.h"
 
 namespace clang {
 namespace spirv {
